feat: create ticketStats for just tickets

+++ src/main/java/com/alexsysSolutions/alexsis/dto/response/dashboard/TicketStatsDto.java
@@ -4,11 +4,12 @@ import lombok.Data;
 
 @Data
 public class TicketStatsDto {
+
     private Long totalTickets;
     private Long totalActiveTickets; // tickets with status OPEN or IN_PROGRESS
     private Long totalResolvedTickets;
     private Long totalClosedTickets;
-    private int avgResolutionTime; // average resolution time in hours (from agents)
+    private Double avgResolutionTime; // average resolution time in hours (from agents)
     private int highPriorityTickets;
     private int totalTicketsToday;
 }
